src/main/python/ai_service/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv

